Change responseHandler location
