+++ src/services/loginService.js
@@ -1,9 +1,23 @@
-import {apiUrl, handleResponse} from 'helpers/config';
+import {apiUrl} from 'helpers/config';
  
 const loginService = {
     login,
     logout
 };
+
+export const handleResponse = (response) => {
+    return response.json().then(data => {
+        if (!response.ok) {
+            if (response.status === 401) {
+                logout();
+            }
+            const error = (data && data.error) || response.statusText;
+            return Promise.reject(error);
+        }
+ 
+        return data;
+    });
+}
  
 function login (email, password) {
     const requestOptions = {
